HTBHF-1863 move handle path request to directory
