+++ src/web/routes/application/middleware/handle-path-request/handle-path-request.test.js
@@ -1,8 +1,8 @@
 const test = require('tape')
 const sinon = require('sinon')
-const { CHECK_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../common/constants')
+const { CHECK_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../../common/constants')
 const { getPathsInSequence, handleRequestForPath, isPathInSequence } = require('./handle-path-request')
-const { states } = require('../common/state-machine')
+const { states } = require('../../common/state-machine')
 
 const steps = [{ path: '/first', next: () => '/second' }, { path: '/second' }]
