+++ src/web/routes/application/middleware/handle-path-request/index.js
@@ -1,0 +1,5 @@
+const { handleRequestForPath } = require('./handle-path-request')
+
+module.exports = {
+  handleRequestForPath
+}
